Refactor page to make the gallery itself into a client component
