+++ gt.config.json
@@ -6,5 +6,5 @@
   "locales": [
     "ja"
   ],
-  "_versionId": "58509547aa0e24502d28455e02f3cd4e798e1dd7dbd9190c8ba9a4483d0a72d3"
+  "_versionId": "0ab02022bcd8988758b31dfe8231eb12714adcdcf146c097cdc74e15efdedd24"
 }
